CI: Fix finding test script
Currently, the run test code tries to find the package source directory
based on the directory name only. This fails for the Go compiler package
because there is more than one directory named "golang".

This uses the full path listed in the "Source:" line of the control file
to find the package source directory.

This also checks for the test script earlier, to avoid installing and
removing ipk files when there is no test script to be run.

Signed-off-by: Jeffery To <jeffery.to@gmail.com>